+++ src/api/sk_txn.h
@@ -30,5 +30,8 @@ struct sk_module_t* sk_txn_current_module(sk_txn_t* txn);
 int sk_txn_is_first_module(sk_txn_t* txn);
 int sk_txn_is_last_module(sk_txn_t* txn);
 
+// sk_txn's alive time from it be created
+unsigned long long sk_txn_alivetime(sk_txn_t* txn);
+
 #endif