+++ src/common/sk_env.c
@@ -30,3 +30,18 @@ sk_thread_env_t* sk_thread_env()
 {
     return pthread_getspecific(sk_env_key);
 }
+
+sk_thread_env_t* sk_thread_env_create(sk_core_t* core,
+                                      sk_engine_t* engine,
+                                      const char* name,
+                                      int idx)
+{
+    sk_thread_env_t* thread_env = calloc(1, sizeof(*thread_env));
+    thread_env->core = core;
+    thread_env->engine = engine;
+
+    snprintf(thread_env->name, SK_ENV_NAME_LEN, "%s", name);
+    thread_env->idx = idx;
+
+    return thread_env;
+}