+++ src/loaders/c_loader.c
@@ -15,9 +15,8 @@ const char* sk_c_module_name(const char* short_name, char* fullname, size_t sz)
 {
     memset(fullname, 0, sz);
 
-    // The full name like: modules/test/mod.so
-    snprintf(fullname, sz, "modules/%s/" SK_MODULE_PREFIX_NAME ".so",
-             short_name);
+    // The full name like: lib/mod.so
+    snprintf(fullname, sz, "lib/" SK_MODULE_PREFIX_NAME "%s.so", short_name);
     return fullname;
 }