+++ src/user/include/skull/api.h
@@ -1,0 +1,16 @@
+#ifndef SKULL_API_H
+#define SKULL_API_H
+
+#ifdef __cplusplus
+extern "C" {
+#endif
+
+#include <skull/logger.h>
+#include <skull/metrics_utils.h>
+
+#ifdef __cplusplus
+}
+#endif
+
+#endif
+