+++ src/user/include/skull/metrics_utils.h
@@ -1,0 +1,19 @@
+#ifndef SKULL_METRICS_UTILS_H
+#define SKULL_METRICS_UTILS_H
+
+#ifdef __cplusplus
+extern "C" {
+#endif
+
+void skull_metric_inc(const char* name, double value);
+double skull_metric_get(const char* name);
+
+typedef void (*skull_metric_each)(const char* name, double value, void* ud);
+void skull_metric_foreach(skull_metric_each, void* ud);
+
+#ifdef __cplusplus
+}
+#endif
+
+#endif
+