+++ src/user/src/logger_private.c
@@ -9,7 +9,7 @@
 
 void skull_log(const char* fmt, ...)
 {
-    sk_logger_t* logger = SK_THREAD_ENV_LOGGER;
+    sk_logger_t* logger = SK_ENV_LOGGER;
 
     va_list ap;
     va_start(ap, fmt);
@@ -50,42 +50,42 @@ bool skull_log_enable_fatal()
 
 const char* skull_log_info_msg(int log_id)
 {
-    sk_log_tpl_t* info_tpl = SK_THREAD_ENV_CORE->info_log_tpl;
+    sk_log_tpl_t* info_tpl = SK_ENV_CORE->info_log_tpl;
     return sk_log_tpl_msg(info_tpl, log_id);
 }
 
 const char* skull_log_warn_msg(int log_id)
 {
-    sk_log_tpl_t* warn_tpl = SK_THREAD_ENV_CORE->warn_log_tpl;
+    sk_log_tpl_t* warn_tpl = SK_ENV_CORE->warn_log_tpl;
     return sk_log_tpl_msg(warn_tpl, log_id);
 }
 
 const char* skull_log_warn_solution(int log_id)
 {
-    sk_log_tpl_t* warn_tpl = SK_THREAD_ENV_CORE->warn_log_tpl;
+    sk_log_tpl_t* warn_tpl = SK_ENV_CORE->warn_log_tpl;
     return sk_log_tpl_solution(warn_tpl, log_id);
 }
 
 const char* skull_log_error_msg(int log_id)
 {
-    sk_log_tpl_t* error_tpl = SK_THREAD_ENV_CORE->error_log_tpl;
+    sk_log_tpl_t* error_tpl = SK_ENV_CORE->error_log_tpl;
     return sk_log_tpl_msg(error_tpl, log_id);
 }
 
 const char* skull_log_error_solution(int log_id)
 {
-    sk_log_tpl_t* error_tpl = SK_THREAD_ENV_CORE->error_log_tpl;
+    sk_log_tpl_t* error_tpl = SK_ENV_CORE->error_log_tpl;
     return sk_log_tpl_solution(error_tpl, log_id);
 }
 
 const char* skull_log_fatal_msg(int log_id)
 {
-    sk_log_tpl_t* fatal_tpl = SK_THREAD_ENV_CORE->fatal_log_tpl;
+    sk_log_tpl_t* fatal_tpl = SK_ENV_CORE->fatal_log_tpl;
     return sk_log_tpl_msg(fatal_tpl, log_id);
 }
 
 const char* skull_log_fatal_solution(int log_id)
 {
-    sk_log_tpl_t* fatal_tpl = SK_THREAD_ENV_CORE->fatal_log_tpl;
+    sk_log_tpl_t* fatal_tpl = SK_ENV_CORE->fatal_log_tpl;
     return sk_log_tpl_solution(fatal_tpl, log_id);
 }